+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # VCF Check
 
+## 2.0.0.1007
+
+### Release Date : 2026-09-11
+
+* Bug Fix: Resolve polling issue caused by stalled task in `SDDC Manager Health Summary` check.
+* Bug Fix: Numberous improvements to `vSAN Health Check` (functionality, UI, and bug fixes).
+* Bug Fix: Improve edge-case error handling in `NSX Manager and Edge Password Expiration`. 
+
 ## 2.0.0.1006
+
+### Release Date : 2026-09-10
 
 * New platform: Powershell Engine with Python UI
 * Aria Operations checks added as a category